Fixed typo
I don't even know if I can add another task, that is Beat tracking. Everything inside an application could be really useful for musicians that wants to improvise on a solo or during a song. If I see that everything gets clear and there won't be any negative results, I can think about to add this task.